rotation stuck fix

+++ 2.91/Welder/utils.py
@@ -132,7 +132,7 @@ def enabledatatransfer(obj):
         if m.type=='CURVE' and obj.scale[1]<1:
             curve=m.object
             if (not curve==None):
-                curve.data.resolution_u=1/obj.scale[2]*3
+                curve.data.resolution_u=int(1/obj.scale[2]*3)
 
 def cleanupWeld(obj):
     if (obj.data.shape_keys!=None):
